requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==4.2
django-crispy-forms==2.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
jsmin==3.0.1
Pillow==9.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
PyJWT==1.7.1
pytz==2023.3
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
webencodings==0.5.1
